Move prend maintenant double au lieu de int, ajout des valeurs de chaque pièce, fonction permettant de retourner valeur d'une pièce, ajout du calcul des moves du Rook

+++ ChessIA/ChessIA/Move.cs
@@ -9,9 +9,9 @@ namespace ChessIA
     class Move
     {
         private Position position;
-        private int value;
+        private double value;
 		
-		public Move(Position position, int value)
+		public Move(Position position, double value)
 		{
 			this.position = position;
 			this.value = value;
@@ -21,7 +21,7 @@ namespace ChessIA
 		{
 			return position;
 		}
-		public int getValue()
+		public double getValue()
 		{
 			return value;
 		}
